Wrote tests for adv commands

+++ include/util.hpp
@@ -62,7 +62,7 @@ inline std::vector<std::string> splitSpaces(const std::string& line)
     return retVec;
 }
 
-inline std::string argumentError(std::string expected, int got)
+inline std::string argumentError(std::string expected, size_t got)
 {
     return std::format("-ERR command expected {} arguments, got {} instead\r\n", expected, got);
 }
